adds test for custom formats and slice serializer
