+++ circle.yml
@@ -13,14 +13,11 @@ general:
 dependencies:
   override:
     - docker info
-    - docker build -t go-swagger/go-swagger:ci-go15-$CIRCLE_BUILD_NUM-$CIRCLE_SHA1 -f Dockerfile.go15.ci .
-    - docker build -t go-swagger/go-swagger:ci-$CIRCLE_BUILD_NUM-$CIRCLE_SHA1 -f Dockerfile.ci .
+     - docker build -t go-swagger/go-swagger:ci-$CIRCLE_BUILD_NUM-$CIRCLE_SHA1 -f Dockerfile.ci .
 
 test:
   override:
     - mkdir -p $CIRCLE_TEST_REPORTS/go
     - mkdir -p $CIRCLE_ARTIFACTS/dist
-    - docker run -v $CIRCLE_ARTIFACTS/dist:/usr/share/dist -v $CIRCLE_TEST_REPORTS/go:/usr/share/testresults -v $CIRCLE_ARTIFACTS:/usr/share/coverage -e CIRCLE_BUILD_NUM=$CIRCLE_BUILD_NUM -e CIRCLE_PROJECT_USERNAME=$CIRCLE_PROJECT_USERNAME -e CIRCLE_PROJECT_REPONAME=$CIRCLE_PROJECT_REPONAME go-swagger/go-swagger:ci-go15-$CIRCLE_BUILD_NUM-$CIRCLE_SHA1 ./hack/build-docker.sh:
-        timeout: 1800 # Fail if command has no output for 30 mins.  Probably doesn't need that long.
     - docker run -v $CIRCLE_ARTIFACTS/dist:/usr/share/dist -v $CIRCLE_TEST_REPORTS/go:/usr/share/testresults -v $CIRCLE_ARTIFACTS:/usr/share/coverage -e CIRCLE_BUILD_NUM=$CIRCLE_BUILD_NUM -e CIRCLE_PROJECT_USERNAME=$CIRCLE_PROJECT_USERNAME -e CIRCLE_PROJECT_REPONAME=$CIRCLE_PROJECT_REPONAME go-swagger/go-swagger:ci-$CIRCLE_BUILD_NUM-$CIRCLE_SHA1 ./hack/build-docker.sh:
         timeout: 1800 # Fail if command has no output for 30 mins.  Probably doesn't need that long.
