+++ cmd/swagger/commands/version_number.go
@@ -1,5 +1,6 @@
 package commands
 
 func init() {
-	Version = "0.5.0-122-g68b544d"
+  Version = "0.5.0-124-gd4c182c"
 }
+
