--- fixtures/templates/additionalpropertiesserializer.gotmpl
@@ -1,1 +1,0 @@
-./additionalpropertiesserializer.gotmpl
