--- fixtures/templates/client/client.gotmpl
@@ -1,1 +1,0 @@
-./client/client.gotmpl
