--- fixtures/templates/client/facade.gotmpl
@@ -1,1 +1,0 @@
-./client/facade.gotmpl
