--- fixtures/templates/client/parameter.gotmpl
@@ -1,1 +1,0 @@
-./client/parameter.gotmpl
