--- fixtures/templates/client/response.gotmpl
@@ -1,1 +1,0 @@
-./client/response.gotmpl
