--- fixtures/templates/docstring.gotmpl
@@ -1,1 +1,0 @@
-./docstring.gotmpl
