--- fixtures/templates/gen_templates.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-echo $1 > $1
