--- fixtures/templates/header.gotmpl
@@ -1,1 +1,0 @@
-./header.gotmpl
