--- fixtures/templates/model.gotmpl
@@ -1,1 +1,0 @@
-./model.gotmpl
