--- fixtures/templates/modelvalidator.gotmpl
@@ -1,1 +1,0 @@
-./modelvalidator.gotmpl
