--- fixtures/templates/schema.gotmpl
@@ -1,1 +1,0 @@
-./schema.gotmpl
