--- fixtures/templates/schemabody.gotmpl
@@ -1,1 +1,0 @@
-./schemabody.gotmpl
