--- fixtures/templates/schematype.gotmpl
@@ -1,1 +1,0 @@
-./schematype.gotmpl
