--- fixtures/templates/schemavalidator.gotmpl
@@ -1,1 +1,0 @@
-./schemavalidator.gotmpl
