--- fixtures/templates/server/builder.gotmpl
@@ -1,1 +1,0 @@
-./server/builder.gotmpl
