--- fixtures/templates/server/configureapi.gotmpl
@@ -1,1 +1,0 @@
-./server/configureapi.gotmpl
