--- fixtures/templates/server/doc.gotmpl
@@ -1,1 +1,0 @@
-./server/doc.gotmpl
