--- fixtures/templates/server/main.gotmpl
@@ -1,1 +1,0 @@
-./server/main.gotmpl
