--- fixtures/templates/server/operation.gotmpl
@@ -1,1 +1,0 @@
-./server/operation.gotmpl
