--- fixtures/templates/server/parameter.gotmpl
@@ -1,1 +1,0 @@
-./server/parameter.gotmpl
