--- fixtures/templates/server/responses.gotmpl
@@ -1,1 +1,0 @@
-./server/responses.gotmpl
