--- fixtures/templates/structfield.gotmpl
@@ -1,1 +1,0 @@
-./structfield.gotmpl
