--- fixtures/templates/swagger_json_embed.gotmpl
@@ -1,1 +1,0 @@
-./swagger_json_embed.gotmpl
