--- fixtures/templates/tuplefield.gotmpl
@@ -1,1 +1,0 @@
-./tuplefield.gotmpl
