--- fixtures/templates/tupleserializer.gotmpl
@@ -1,1 +1,0 @@
-./tupleserializer.gotmpl
