--- fixtures/templates/validation/customformat.gotmpl
@@ -1,1 +1,0 @@
-./validation/customformat.gotmpl
