--- fixtures/templates/validation/primitive.gotmpl
@@ -1,1 +1,0 @@
-./validation/primitive.gotmpl
