--- fixtures/templates/validation/structfield.gotmpl
@@ -1,1 +1,0 @@
-./validation/structfield.gotmpl
