+++ generator/bindata.go
@@ -216,7 +216,7 @@ func templatesDocstringGotmpl() (*asset, error) {
 	return a, nil
 }
 
-var _templatesHeaderGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\x7c\x91\xc1\x4e\xeb\x30\x10\x45\xd7\xcf\x5f\x31\x8a\xfa\x24\x90\x68\xbc\x47\x62\x07\x0b\x76\x2c\xf8\x01\xb7\x19\x3b\xd3\x3a\x76\xb0\xc7\x0d\x91\x95\x7f\xc7\xad\x09\x52\x8b\xc4\xee\x46\xf7\xcc\x99\x64\x32\xaa\xfd\x51\x19\x84\x9c\xdb\xb7\x1a\x97\x45\x08\x29\xe1\xbd\xa7\x08\x9a\x2c\xc2\xa4\x22\x18\x74\x18\x14\x63\x07\xbb\x19\xb8\x47\x88\x93\x32\x06\x03\xb0\xf7\xb6\x3d\xf3\x2f\x1d\x31\x39\x53\xca\x75\x6e\x20\xd3\x33\x8c\xc1\x9f\x10\x74\xe2\x8b\xaa\x47\x07\xb3\x4f\x10\x70\x1b\x92\xbb\x32\xad\x2b\x60\xef\x87\x41\xb9\x4e\x88\x9c\x49\x83\x0f\xd0\xbe\x0e\xa3\x0f\x1c\xa1\x7d\x46\xad\x92\xe5\xf5\x79\x59\xe8\x92\xe0\x4e\x00\x44\x0e\x7a\x60\x80\xc6\x10\xf7\x69\xd7\x16\x8d\x34\x7e\xeb\x47\x74\x6a\x24\x59\xeb\xa6\x80\x07\x8b\x9f\x65\xdf\x15\x38\x28\xb2\x21\x49\x54\x71\x3e\x44\xef\x64\x65\x1a\xf1\xef\x30\x05\xe2\x42\xff\x09\x57\xa6\xd0\x3b\xe2\x88\x37\xef\x30\x91\xb5\x5a\xd6\xa6\x11\x65\x7f\xce\x41\xb9\x72\xf2\xdf\x5f\x93\x73\x39\x17\x39\xd6\xd0\xfc\xff\x68\xa0\x2d\xbf\xe2\x8c\xa3\xeb\xbe\x53\x1d\xdc\x1c\x71\x7e\x80\xcd\x49\xd9\x84\xf0\xf8\xf4\x73\x9f\x8b\xe0\x5c\x16\x15\xdc\xb8\x2a\x7d\x25\xbc\x17\x6b\xfa\x0a\x00\x00\xff\xff\xab\xb1\x90\x52\x05\x02\x00\x00")
+var _templatesHeaderGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\x7c\x91\xc1\x6e\xe3\x20\x14\x45\xd7\xc3\x57\x5c\x59\x19\x69\x46\x9a\xd8\xfb\x91\xba\x6b\x17\xdd\x75\xd1\x1f\x20\xf1\x33\x26\xc1\xe0\xc2\x73\x5c\x0b\xf9\xdf\x8b\x4d\x5d\x29\xa9\xd2\xdd\x45\xf7\x70\x78\x7a\xf4\xf2\x78\x96\x8a\x10\x63\xf9\x92\xe3\x3c\x0b\x51\x55\x78\x6d\x75\x40\xa3\x0d\x61\x94\x01\x8a\x2c\x79\xc9\x54\xe3\x30\x81\x5b\x42\x18\xa5\x52\xe4\xc1\xce\x99\x72\xe1\x9f\x6a\xcd\xda\xaa\x54\x6e\xf7\x3a\xad\x5a\x46\xef\xdd\x85\xd0\x0c\xbc\xaa\x5a\xb2\x98\xdc\x00\x4f\x7b\x3f\xd8\x2b\xd3\xf6\x04\x8e\xae\xeb\xa4\xad\x85\x88\x51\x37\x70\x1e\xe5\x73\xd7\x3b\xcf\x01\xe5\x23\x35\x72\x30\xbc\x9d\xe7\x59\xaf\x09\x7f\x04\x40\xde\x3b\x1f\x80\x42\x69\x6e\x87\x43\x99\x34\x95\x72\x7b\xd7\x93\x95\xbd\xae\x72\x5d\x24\x30\xb0\x6f\x3a\xbe\x0b\xe6\x7a\x05\xd3\x64\xc0\x7d\x30\xd5\x0b\x76\x32\xf4\x9e\xe6\xbf\xc2\x3a\xa9\x8d\x1f\x2a\x92\x61\x3a\x05\x67\xab\xcc\x14\xe2\xd7\x69\xf4\x9a\x13\xfd\x23\x9c\x99\x44\x1f\x34\x07\xba\x19\x75\xd4\xc6\x34\x55\x6e\x0a\x91\xde\x8f\xd1\x4b\x9b\xbe\xf0\xfb\x76\x62\x4c\xeb\xd7\x96\x1b\x14\xbf\xdf\x0a\x94\xe9\x6b\x17\x9c\x6c\xfd\x99\xf2\xc5\xdd\x99\xa6\x7f\xd8\x5d\xa4\x19\x08\xff\x1f\xbe\xf6\xbd\x0a\x96\x32\xa9\x70\xe3\xca\xf4\x95\xf0\xaf\xd8\xd2\x47\x00\x00\x00\xff\xff\x24\x48\xae\xf3\x55\x02\x00\x00")
 
 func templatesHeaderGotmplBytes() ([]byte, error) {
 	return bindataRead(
@@ -231,7 +231,7 @@ func templatesHeaderGotmpl() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "templates/header.gotmpl", size: 517, mode: os.FileMode(420), modTime: time.Unix(1465223821, 0)}
+	info := bindataFileInfo{name: "templates/header.gotmpl", size: 597, mode: os.FileMode(420), modTime: time.Unix(1465223932, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -276,7 +276,7 @@ func templatesModelvalidatorGotmpl() (*asset, error) {
 	return a, nil
 }
 
-var _templatesSchemaGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\xec\x58\x4b\x6f\xe3\xb6\x13\xbf\xfb\x53\xcc\xdf\xf0\x7f\x61\x05\x86\x53\x04\x3d\xa5\xc8\x21\xd9\xed\x23\x97\x6e\x51\xa7\x7b\x09\x16\x5d\x5a\xa6\x62\x66\x25\x51\x21\xa9\x64\xdd\x20\xdf\xbd\x33\x7c\xe8\x65\xc9\x8f\xed\x63\x51\xb4\x80\x0f\x32\x39\x9c\xf9\xf1\x37\xa3\x79\xe8\xf9\x19\x44\x02\x2c\x5f\xc1\xfc\x5a\x5f\x31\xcd\x6f\x36\x05\xa7\xe7\x6f\x3f\x15\x52\x19\xbe\x82\x97\x17\x43\x4b\x28\x58\x30\x1d\xb3\x54\xfc\x86\xfb\x3f\xb2\x8c\xe3\x0e\x88\xdc\x70\x95\xb0\x18\xf7\x47\x80\x32\x5e\xd7\x34\x97\x86\x94\x5c\x87\xed\x08\xa6\x52\xc1\xfc\x67\xfe\x50\x0a\x85\x4a\xe7\x3f\x30\xfd\x0e\x75\xad\x98\x11\x32\xd7\x11\xea\x52\x65\x6e\x44\xc6\xe7\x7e\x99\x2d\x53\x8e\x36\x79\x4e\x08\xac\x6e\x50\x2c\xbf\x43\xdb\x97\x69\xfa\x36\xa9\x16\xd1\x22\x1a\xba\xcc\x65\xbe\xc9\x64\xa9\x71\xbd\x96\xfc\x49\xc9\x82\x2b\x23\xb8\x6e\x8a\x4f\x50\xfe\xa6\x2c\x52\xee\x64\x0d\xcf\x8a\x94\x19\x0e\x63\x43\x8b\x89\xe0\xe9\xea\x9a\x30\x8f\x61\xee\x24\x78\xaa\x9d\x6c\x2d\xaa\x8d\x2a\x63\xd3\x27\xdb\xc0\xeb\x9e\x3d\x46\xbc\xf0\xe5\x6a\x25\xe8\xba\x2c\x6d\x01\xf3\x02\x03\xbb\xa7\x27\xd0\x02\xb9\x92\x31\x1a\x17\xf9\xdd\x78\xf0\x48\x4b\xbe\x70\x3b\x9b\x9a\xed\x37\x32\x5e\xec\xd2\x80\x6e\x3d\x39\x75\x37\x68\x78\xbc\x4f\x32\x84\xc1\x34\x82\x8c\x15\xb7\x0e\xd7\xfb\x96\x79\x1d\xaf\x79\xc6\x28\xa8\x86\xf1\x92\x29\xe4\x2a\xf0\xd7\xf4\x6c\x7d\xe2\x1a\x75\x1e\xce\x47\x90\xfe\x2c\x2a\xec\xe1\x7d\x2c\x58\xa1\x06\x01\xb7\x07\xdd\x3b\xe0\x6a\x06\x88\x7f\x76\x41\xe6\xfe\xcc\xbf\x97\xf6\x3d\x1c\x08\x29\xfb\xbc\x15\xe3\x5f\x20\xc4\x3b\xde\xfa\x2f\xc6\x07\xf1\x76\x32\x42\xd3\xa7\xff\x9a\x38\x7f\x19\x8d\x4e\x4f\xe1\x97\x3c\x63\x4a\xaf\x59\xda\x5b\x51\x16\xa9\xc0\x62\x52\x06\x19\x0d\x85\x4c\x31\xb1\xab\x62\x2d\x62\xd0\xb4\xa9\x41\x26\xfd\xd5\x68\x94\x94\x79\x7c\x88\xfe\xa9\xe2\x6c\xc5\x15\x08\x89\x15\x89\x9e\x66\x10\x63\x15\x2a\x33\x5c\x0b\x65\xe8\xb5\x5f\xc0\xca\x65\xaf\xbc\xad\x6a\x06\x5c\x29\x89\x02\x54\xfa\x1e\x99\xc2\x97\x88\x67\x3c\x37\x1a\x39\xba\x7d\xbf\xdc\x18\x8e\xeb\xe8\x5c\x94\x82\xf3\x8b\xca\x42\xd0\xec\x41\xcc\xe0\x55\x38\x17\x7d\x63\x65\xff\x77\x01\xb9\x48\xad\x56\x00\xc5\x4d\xa9\x72\x5a\xb0\xe6\x70\x0d\x59\x74\xe6\x14\xd7\x65\x6a\x60\x00\x1d\x0a\x25\x58\x72\x7f\x9d\x05\x58\x84\xc1\xe5\x8c\x0a\xa7\x33\x21\x97\xf7\xb3\x00\xb2\xdc\x49\xde\xd4\x9f\xac\xe9\x8a\xac\x06\x7f\xc9\x16\xf0\x3e\xe8\x04\xde\xed\x58\xe4\x17\xc0\x8a\x02\x63\x63\xea\xfe\xcf\x08\x49\x34\x72\x42\xfe\x30\x84\x2d\xd4\x42\xf1\xb3\x3f\x80\x86\x62\xe7\xb3\x23\xe6\xc8\x60\xd9\x1f\x2a\x78\x85\x27\x0e\x39\xc7\x36\xc8\x48\x20\xed\x60\xd6\x42\x83\x79\xc2\xd0\x9c\x81\x96\x90\x08\xa5\x0d\xf5\x56\x12\x18\x2c\xcb\x24\xe1\xc4\x1e\x35\x45\x95\xa3\x84\x2c\x8d\x48\x2d\x22\xec\x87\x3c\xc6\x68\xd4\xef\x8b\xbe\x20\xaa\x29\xde\xe3\x73\x67\xb6\x76\x38\x7a\xc1\xb2\x76\xc0\x31\x70\xaf\xc1\x1f\x25\x0c\x19\xa0\x2b\x93\x2a\xcc\x44\xfc\xe9\xca\x32\x62\x2d\x44\x6e\xfb\x6c\x78\xdf\x11\x6e\xd6\xdc\xb3\x4a\xe6\x1d\xdf\xf8\xb3\xe4\x13\xf5\xc8\x39\x37\xf1\xda\xca\x3d\xb2\xb4\xe4\x94\x64\xe8\x0f\x15\xe3\x37\x42\xc7\x4a\x64\x22\x67\x46\xaa\xef\xa8\x20\x52\x9c\x85\x2c\x3b\xf7\xaf\xe3\x1d\x37\xb6\x66\xbb\xba\x09\xcf\x9d\x88\xeb\x57\xe2\x52\x3a\x7c\xb8\xd7\x32\x3f\xa7\x03\xf8\xd7\x24\x30\xfe\xff\xc3\x78\xe0\xc8\x07\xeb\xbb\x1d\x69\x05\xe9\xc0\x9c\xe2\xd1\x1c\x91\x52\x6a\x95\x8f\xae\x6e\xf0\xaa\x5d\x77\xb5\x63\x7a\x10\xbe\x19\x8c\x97\x72\xb5\x19\xcf\x02\x21\xf3\x03\x78\x38\x02\x26\x3a\xf3\xa6\xe9\xa4\x61\x07\xa1\x5f\x4b\xed\x5e\xb2\x15\xc7\x39\x04\xf7\x39\x3c\x61\x2e\x40\x37\x93\xa3\x70\x3d\xc6\x00\xc0\x2a\x46\x23\x4b\x15\xce\xd6\xed\x36\x7a\xe9\x05\x44\x8b\xfa\x49\x50\x68\x1c\x71\x1d\xe7\x7c\x97\x6c\x27\x1f\x67\x30\x79\x24\x5a\x9b\xb2\xa1\x27\x00\x88\x71\xe6\x82\x0e\xb3\x93\x8f\xb8\x7b\xee\xd3\x68\x23\xd5\xa3\x18\xaa\xf2\x07\xf7\x05\xc1\x19\x46\x81\x3b\xd7\xc7\xee\x50\x82\x0e\x29\xba\xda\x7d\xd5\xcc\xc0\xb4\xde\xec\x5e\x1a\x69\x24\x68\x91\xca\xbe\x83\xd3\xaf\xcf\x10\xc0\x58\xe4\x36\x98\x76\x78\xc9\x3a\xf2\x1c\xf0\xda\xc7\x45\xcc\x08\x13\x51\xb3\x71\x25\x36\x68\xc4\xbc\xd6\xaf\x25\x36\x27\xfc\xd3\xdb\xe5\x3d\x8f\xed\x14\xea\x3a\x61\x9a\x12\x77\x36\xa7\x3e\xf5\x84\x69\x17\x97\xfc\x14\xdb\x18\x85\xe9\x1e\x5e\xae\x65\x7c\x3b\x83\x55\x34\xb5\xfa\xbe\x6e\xe3\x74\x45\xaf\x8a\x6d\xac\x1b\x97\xd9\x31\x6f\xb7\x9b\x79\x94\x5b\x34\xfb\xaf\xa0\xc6\xb9\xa7\x1a\x8e\x17\xe5\xb2\x9e\x26\xba\x43\xf4\xd0\xb4\x5c\x9d\xae\x3e\x0a\xbc\xbc\xd8\xbc\x8f\x69\x60\x82\x99\x21\xe6\xe2\x91\x2b\xc2\x45\x1d\x64\x0b\xed\x64\xee\x96\xa3\x9e\x4b\xd8\x1e\x72\xb8\x83\x24\x8c\x55\x57\xcc\x1f\x50\x55\x8f\xf7\x03\x1b\x3e\xf2\xba\x2f\x4f\xfb\xc8\x3b\x9b\x28\x9a\xee\xaa\x8f\xb5\xef\x81\x5b\x18\x79\x31\x3e\x35\xe1\x5a\x93\x61\xd2\xf1\x0d\xc3\x31\x14\x2c\xb8\xe9\x65\x01\xe3\x7e\x37\x0f\x11\xd4\x4c\xe4\x7c\x37\x13\xc7\xdc\x05\x6c\x82\xaf\x6f\xb4\x3d\x96\xfc\xd3\x3f\x5d\xfc\x6d\x43\x5d\xeb\xb3\x45\xe3\xa5\xf9\xd2\xd3\xdc\x5f\x34\xcb\xf5\x06\x48\xf7\xf8\x82\x2b\x61\x0d\xab\xf0\xc1\x60\xc7\x37\x41\xbf\x14\x52\xcc\x9e\xaf\x84\xf6\x7f\x2d\xeb\x10\x75\xed\xfb\xe6\x45\xaa\x4e\x5a\x45\x14\x77\x06\xa6\x29\xcf\x7d\xf2\x8b\xe0\xab\xe3\x55\x10\xe0\xa9\x2b\x33\xd5\x3d\x6c\x8a\x35\xd8\x49\x64\x84\x09\xfb\x13\x0f\x99\x57\x8d\x94\x76\x0d\x27\xea\x59\x97\x19\xcb\xb7\x27\x10\x4c\x28\xdd\x7c\xd2\xac\x68\x55\x01\xdb\x2a\x6d\x03\x81\x79\xd2\x79\x85\xff\x94\x42\x16\x55\x17\x9b\xe2\x64\x99\x31\x1c\x20\x31\x86\x93\xcc\x20\xf4\x3b\x81\x8f\x9b\xc8\x35\x00\x36\x71\xd5\x6d\xc1\xa8\x27\xb3\x8c\x7e\x0f\x00\x00\xff\xff\xac\xa2\x56\xc2\x78\x16\x00\x00")
+var _templatesSchemaGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\xec\x58\x4b\x6f\xdb\xc6\x13\xbf\xeb\x53\xcc\x5f\xd0\x3f\x10\x03\x41\x2e\x82\x9e\x5c\xe4\x10\x27\x7d\xf8\xd2\x14\x95\x9b\x8b\x11\x34\x2b\x6a\x68\xad\x43\x72\xe9\xdd\xa5\x1d\xd5\xf0\x77\xef\xcc\xee\xf2\x25\x91\x7a\xa4\x8f\xa0\x68\x01\x1d\xa8\xdd\xd9\xd9\xdf\xfc\x66\x38\x0f\x3e\x3e\x82\x4c\x40\xe4\x2b\x98\x5f\x9a\x0b\x61\xf0\x6a\x53\x20\x3f\x7f\xfb\xa9\x50\xda\xe2\x0a\x9e\x9e\x2c\x2f\x91\x60\x21\x4c\x2c\x52\xf9\x1b\xed\xff\x28\x32\xa4\x1d\x90\xb9\x45\x9d\x88\x98\xf6\x47\x40\x32\x41\xd7\x34\x57\x96\x95\x5c\x56\xdb\x11\x4c\x95\x86\xf9\xcf\x78\x57\x4a\x4d\x4a\xe7\x3f\x08\xf3\x8e\x74\xad\x84\x95\x2a\x37\x11\xe9\xd2\x65\x6e\x65\x86\xf3\xb0\x2c\x96\x29\xd2\x9d\x98\x33\x02\xa7\x1b\xb4\xc8\x6f\xe8\xee\x57\x69\xfa\x36\xa9\x17\xe9\x46\xba\xe8\x55\xae\xf2\x4d\xa6\x4a\x43\xeb\x8d\xe4\x4f\x5a\x15\xa8\xad\x44\xd3\x16\x9f\x90\xfc\x55\x59\xa4\xe8\x65\x2d\x66\x45\x2a\x2c\xc2\xd8\xf2\x62\x22\x31\x5d\x5d\x32\xe6\x31\xcc\xbd\x04\xa6\xc6\xcb\x36\xa2\xc6\xea\x32\xb6\x7d\xb2\x2d\xbc\xfe\x39\x60\x24\x83\x5f\xad\x56\x92\xcd\x15\x69\x07\x58\x10\x18\xd8\x3d\x7b\x0e\x1d\x90\x2b\x15\xd3\xe5\x32\xbf\x19\x0f\x1e\xe9\xc8\x17\x7e\x67\xd3\xb0\xfd\x46\xc5\x8b\x7d\x1a\xc8\xad\xcf\xcf\xbc\x05\x2d\x8f\xf7\x49\x56\x61\x30\x8d\x20\x13\xc5\xb5\xc7\xf5\xbe\x73\xbd\x89\xd7\x98\x09\x0e\xaa\x61\xbc\x7c\x15\x71\x55\xf1\xd7\xf6\x6c\x73\xe2\x92\x74\x1e\xcf\x47\x25\xfd\x59\x54\xb8\xc3\x87\x58\x70\x42\x2d\x02\xae\x8f\xb2\xbb\xc2\xd5\x0e\x90\xf0\xec\x83\xcc\xff\x99\x7f\xaf\xdc\x7b\x38\x10\x52\xee\x79\x27\xc6\xbf\x40\x88\x6f\x79\xeb\xbf\x18\x1f\xc4\xbb\x95\x11\xda\x3e\xfd\xd7\xc4\xf9\xd3\x68\x74\x76\x06\xbf\xe4\x99\xd0\x66\x2d\xd2\xde\x8a\xb2\x48\x25\x15\x93\xb2\x92\x31\x50\xa8\x94\x12\xbb\x2e\xd6\x32\x06\xc3\x9b\x06\x54\xd2\x5f\x8d\x46\x49\x99\xc7\xc7\xe8\x9f\x6a\x14\x2b\xd4\x20\x15\x55\x24\x7e\x9a\x41\x4c\x55\xa8\xcc\x68\xad\x2a\x43\xaf\xc3\x02\x55\x2e\x67\xf2\xae\xaa\x19\xa0\xd6\x8a\x04\xb8\xf4\xdd\x0b\x4d\x2f\x11\x66\x98\x5b\x43\x1c\x5d\xbf\x5f\x6e\x2c\xd2\x3a\x39\x97\xa4\xe0\xfc\x65\x7d\x43\xa5\x39\x80\x98\xc1\xb3\xea\x5c\xf4\x8d\x93\xfd\xdf\x4b\xc8\x65\xea\xb4\x02\x68\xb4\xa5\xce\x79\xc1\x5d\x47\x6b\xc4\xa2\xbf\x4e\xa3\x29\x53\x0b\x03\xe8\x48\x28\xa1\x92\xfb\xeb\xac\x82\xc5\x18\x7c\xce\xa8\x71\xfa\x2b\xd4\xf2\x76\x56\x81\x2c\xf7\x92\x37\x0d\x27\x1b\xba\x22\xa7\x21\x18\xd9\x01\xde\x07\x9d\xc1\xfb\x1d\x87\xfc\x25\x88\xa2\xa0\xd8\x98\xfa\xff\x33\x46\x12\x8d\xbc\x50\x38\x0c\xd5\x16\x69\xe1\xf8\x39\x1c\x40\x43\xb1\xf3\xd9\x11\x73\x62\xb0\x1c\x0e\x15\x32\xe1\x01\x21\x47\x6a\x83\xac\x02\xd6\x0e\x76\x2d\x0d\xd8\x07\x0a\xcd\x19\x18\x05\x89\xd4\xc6\x72\x6f\xa5\x40\xc0\xb2\x4c\x12\x64\xf6\xb8\x29\xaa\x1d\x25\x55\x69\x65\xea\x10\x51\x3f\x14\x30\x46\xa3\x7e\x5f\xf4\x05\x51\x43\xf1\x01\x9f\xfb\x6b\x1b\x87\x93\x17\x1c\x6b\x47\x1c\x03\xff\x1a\xfc\x51\xc2\x88\x01\x36\x99\x55\x51\x26\xc2\x87\x0b\xc7\x88\xbb\x21\xf2\xdb\x2f\x86\xf7\x3d\xe1\x76\x8d\x81\x55\xbe\xde\xf3\x4d\x3f\x47\x3e\x53\x4f\x9c\xa3\x8d\xd7\x4e\xee\x5e\xa4\x25\x72\x92\xe1\x3f\x5c\x8c\xdf\x48\x13\x6b\x99\xc9\x5c\x58\xa5\xbf\xe3\x82\xc8\x71\x56\x65\xd9\x79\x78\x1d\x6f\xd0\xba\x9a\xed\xeb\x26\x3c\x6e\x45\x5c\xbf\x12\x9f\xd2\xe1\xc3\xad\x51\xf9\x39\x1f\xa0\xbf\x36\x81\xf1\xff\xef\xc6\x03\x47\x3e\x38\xdf\xed\x49\x2b\x44\x07\xe5\x94\x80\xe6\x84\x94\xd2\xa8\xbc\xf7\x75\x03\xeb\x76\xdd\xd7\x8e\xe9\x51\xf8\x66\x30\x5e\xaa\xd5\x66\x3c\xab\x08\x99\x1f\xc1\xc3\x09\x30\xc9\x99\x57\x6d\x27\x0d\x3b\x88\xfc\x5a\x1a\xff\x92\xad\x90\xe6\x10\xda\x47\x78\xa0\x5c\x40\x6e\x66\x47\xd1\x7a\x4c\x01\x40\x55\x8c\x47\x96\x3a\x9c\x9d\xdb\x5d\xf4\xf2\x0b\x48\x37\x9a\x07\xc9\xa1\x71\x82\x39\xde\xf9\x3e\xd9\x4e\x3e\xce\x60\x72\xcf\xb4\xb6\x65\xab\x9e\x00\x20\xa6\x99\x0b\xb6\x98\x9d\x7c\xa4\xdd\xf3\x90\x46\x5b\xa9\x9e\xc4\x48\x55\x38\x78\x28\x08\x5e\x50\x14\xf8\x73\x7d\xec\x0e\x25\xe8\x2a\x45\xd7\xbb\xcf\xda\x19\x98\xd7\xdb\xdd\x4b\x2b\x8d\x54\x5a\x94\x76\xef\xe0\xf4\xeb\x17\x04\x60\x2c\x73\x17\x4c\x7b\xbc\xe4\x1c\x79\x0e\x64\xf6\x69\x11\x33\xa2\x44\xd4\x6e\x5c\x99\x0d\x1e\x31\x2f\xcd\x6b\x45\xcd\x09\x7e\x7a\xbb\xbc\xc5\xd8\x4d\xa1\xbe\x13\xe6\x29\x71\x6f\x73\x1a\x52\x4f\x35\xed\xd2\x52\x98\x62\x5b\xa3\x30\xdb\x11\xe4\x3a\x97\xef\x66\xb0\x9a\xa6\x4e\xdf\xb7\xdd\x38\x5d\xf0\xab\xe2\x1a\xeb\x96\x31\x7b\xe6\xed\x6e\x33\x4f\x72\x8b\x76\xff\x55\xa9\xf1\xee\xa9\x87\xe3\x45\xb9\x6c\xa6\x89\xed\x21\x7a\x68\x5a\xae\x4f\xd7\x1f\x05\x9e\x9e\x5c\xde\xa7\x34\x30\xa1\xcc\x10\xa3\xbc\x47\xcd\xb8\xb8\x83\xec\xa0\x9d\xcc\xfd\x72\xd4\x63\x84\xeb\x21\x87\x3b\x48\xc6\x58\x77\xc5\x78\x47\xaa\x7a\xbc\x5f\xb1\x11\x22\x6f\xfb\xe5\xe9\x1e\x79\xe7\x12\x45\xdb\x5d\xcd\xb1\xae\x1d\xb4\x45\x91\x17\xd3\x53\x1b\xae\xbb\xb2\x9a\x74\x42\xc3\x70\x0a\x05\x0b\xb4\xbd\x2c\x50\xdc\xef\xe7\x21\x82\x86\x89\x1c\xf7\x33\x71\x8a\x2d\xe0\x12\x7c\x63\xd1\xee\x58\xf2\x4f\xff\x74\xf1\xb7\x0d\x75\x9d\xcf\x16\xad\x97\xe6\x4b\x4f\x73\x7f\xd1\x2c\xd7\x1b\x20\xcd\xf1\x4c\xad\x30\x5d\xa0\x96\xee\x5e\x5d\x7d\x2f\xd8\xf3\x49\x30\x2c\x55\x19\xe6\xc0\x47\x42\xf7\xbf\x91\xf5\x80\xb6\xd1\x87\xde\x45\xe9\xad\xac\x4a\x28\x6e\x2c\x4c\x53\xcc\x43\xee\x8b\xe0\xab\xd3\x55\x30\xe0\xa9\xaf\x32\xb5\x1d\x2e\xc3\x5a\x6a\x24\x32\xc6\x44\xed\x49\x80\x8c\x75\x1f\x65\x7c\xbf\x49\x7a\xd6\x65\x26\xf2\xdd\x01\x84\xf2\xc9\x76\x3a\x69\x17\xb4\xba\x7e\xed\x54\xb6\x81\xb8\x7c\xbe\xf5\x06\xff\x29\x75\x2c\xaa\x0d\x9b\xd2\x60\x99\x09\x9a\x1f\x29\x84\x93\xcc\x12\xf4\x1b\x49\x8f\x9b\xc8\xd7\x7f\x97\xb7\x9a\xae\x60\xd4\x93\x58\x46\xbf\x07\x00\x00\xff\xff\x04\xb9\x7f\x73\x77\x16\x00\x00")
 
 func templatesSchemaGotmplBytes() ([]byte, error) {
 	return bindataRead(
@@ -291,7 +291,7 @@ func templatesSchemaGotmpl() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "templates/schema.gotmpl", size: 5752, mode: os.FileMode(420), modTime: time.Unix(1465223887, 0)}
+	info := bindataFileInfo{name: "templates/schema.gotmpl", size: 5751, mode: os.FileMode(420), modTime: time.Unix(1465223932, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -316,7 +316,7 @@ func templatesSchemabodyGotmpl() (*asset, error) {
 	return a, nil
 }
 
-var _templatesSchemaserializerGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\xcc\x58\xd9\x72\xdb\x36\x17\xbe\x36\x9f\x02\xe1\xf8\xff\x2d\xba\x2e\xd3\xf6\xd2\x33\xb9\x88\x97\x4e\xdd\x49\xed\x4e\xe4\x34\x33\x6d\x3a\x1e\x98\x3c\xb2\x90\x90\xa0\x0a\x80\x95\x55\x85\xef\xde\x03\x80\xfb\xa6\xc5\x71\xdb\x1b\x9b\x04\x71\x96\xef\x3b\x0b\x0e\xb4\x5e\x93\x10\x66\x8c\x03\x71\xe3\x24\x84\x68\x0a\x82\xd1\x88\xfd\x05\xc2\x25\x59\xe6\xbc\x7c\x49\x7e\xa2\x42\xce\x69\xf4\xe3\xf4\xe6\x9a\xc4\xf6\x59\x12\x35\x67\x92\x24\xf7\x1f\x21\x50\x64\xc9\xd4\x9c\xd0\x30\x64\x8a\x25\x9c\x46\x64\x21\x92\x05\x08\xc5\x40\x12\xc6\x55\x42\x28\x31\xb2\x76\xb7\x33\x4b\x79\x40\x26\xeb\xb5\xff\x16\x02\x60\x7f\x82\xb8\xa6\x31\x64\x19\x41\x47\x16\x54\x06\xc6\x36\xf1\xf5\x22\xda\xf7\xea\xd6\x27\x1e\x99\xfc\xf6\xfb\xfd\x4a\xc1\x09\x01\x21\x12\xe1\x91\xb5\x73\x90\xa4\x8a\x9c\xbe\x22\x1f\x97\x82\x29\x10\xfe\x7b\xf3\x6f\x9d\x39\x07\xa8\xb0\x61\x03\xd5\xf9\xb9\xb6\x4b\x2a\x57\x46\xe3\xff\x51\xda\x73\x0e\x04\xa8\x54\x70\x82\x2f\xfe\x59\xca\xa2\xf0\x0c\x6d\xc8\x89\xe7\x64\x8e\x26\xe0\x1d\x8f\x6b\x14\xa4\x7c\x57\x12\x66\x22\x89\x0d\x05\x83\xd8\x8f\x07\xc0\x37\x2c\x4f\x42\xaa\x28\xb1\x04\x78\x96\x00\xc4\x4f\x90\x63\x83\x3f\x82\x47\x84\xff\x46\xff\x5d\x5f\xe0\xce\x53\xa2\xf7\x0f\xf0\x50\x2a\xae\x98\x60\xbc\x22\x82\x71\xff\x52\xeb\x2f\x29\x68\xf1\xf6\x9c\x79\x30\xc4\x45\x3b\x74\x3a\xee\xc7\xcd\xa8\x7b\x86\x0f\x94\x57\x10\x2f\x22\xaa\x30\xa7\x65\x30\x87\x98\xd6\x93\xda\xd7\x69\xdd\x0a\x6c\x89\xeb\x5f\x09\x6e\x09\x09\x23\x79\x5c\x8f\xe3\x20\x9e\x0b\x90\x5d\x44\xb8\x0f\x78\xa8\x9f\x6b\x35\xdd\xc5\x6f\xbf\xb3\x19\xc1\xf4\xf1\xaf\xe4\x79\x82\xaa\xe1\xf1\xc6\x02\xc5\x85\xdb\x14\xdf\xf5\xc3\xeb\x12\xef\xcf\x15\xdc\x2c\xeb\x71\xc7\x0a\x77\x48\xd6\x0e\x45\x12\xb4\x2d\xad\x4e\x08\xba\xea\x97\x9f\x46\x2c\x80\x0d\xe2\xe7\xa9\x54\x49\xfc\x7d\x22\x62\xaa\x30\xd4\x03\x8a\x94\x98\xc5\xc3\x8e\x18\x19\x5f\x53\x7d\xc5\x8b\x75\xc3\xd8\x08\x75\x4d\xaa\x9f\x8d\xbc\x9e\x88\xee\x46\xdf\x46\x05\x3b\x10\x38\xa2\xab\xa2\xf0\x06\xeb\x6f\x4b\x0e\xbb\x61\xc9\xc6\xd3\xb5\xc7\x8b\x0d\x12\xed\x0c\xda\x62\xfb\x76\xfa\x6d\x07\xf8\x85\x46\xe9\x96\x02\xc3\x51\xb5\xfb\x3a\xbd\xf8\xee\x8e\xa7\x11\x36\x9c\x57\xe4\xf8\x9e\x29\x09\xca\xbf\x86\x25\xf6\x11\xc2\x93\x5a\xee\xe8\x28\x78\x03\xf2\x28\x63\x1c\xdc\x4e\x07\xa6\x03\xf6\xf7\x2b\x79\x8d\x56\x27\x5e\x7e\x84\xf8\xd3\x4f\x6c\x81\xed\x9e\x10\x7b\x04\x60\x3f\xc1\xc5\x0b\x88\x58\x3c\x39\x5a\x1f\x79\xce\x0c\x33\xfe\x85\x91\xcb\x17\xb3\x23\x2b\xfb\x09\x56\xfa\xfc\xc1\x4f\xef\xb8\xa4\x33\xc0\xe0\x31\xfe\x60\x54\xe1\xda\x7b\xca\xd5\x79\x12\x25\xdc\x2e\x74\x4d\x13\x22\xb1\xbb\x06\x73\xa3\xc7\x2e\xa0\xdb\x87\xba\xc3\xbe\x01\x73\xb2\x4d\x9a\x18\x3c\x9b\x6b\x82\xf2\x07\x20\x87\x8c\x87\xf0\x78\x62\xf7\xeb\xcd\x7e\xa3\xdc\x8c\xba\x80\x62\xde\xea\x1e\x8c\x7e\xa9\x19\x71\xff\xf7\x87\x6b\xf7\x17\xad\xf8\xd4\x6c\xb3\x76\x3b\xec\x36\x9a\x77\x43\x0c\xc9\x2e\x3c\xf5\x7f\x05\x91\x14\xf6\x06\x14\xd5\xc2\xe4\x4f\x41\xe9\xe0\x58\xef\x4d\x50\x46\x05\x4d\x7e\x0c\x08\xd5\xeb\x4f\x17\xfa\x40\xe3\xd9\x1a\x62\xaf\x82\x26\xe4\xfe\x2d\x65\x5b\x9d\x6c\x40\xd3\xa5\xa1\x08\x76\x0f\x88\x2b\xec\x50\xda\x7f\xf2\x35\xf9\xb6\x84\xba\x13\x5d\xfb\x29\x1f\xa6\xb5\x10\xda\x87\x51\x23\x3b\x48\xa6\xfd\xfa\x45\x78\xdc\x93\xa1\x16\x76\xf3\x86\x7d\x8d\xa6\x91\x2a\x8a\x24\x6f\x15\xa8\x33\x15\x12\xb5\xe6\x5e\xda\xbd\xf5\x3e\x52\x2b\xff\x38\xa6\xf9\x52\x90\x70\xc5\x78\x0a\x8e\x95\x68\x95\xfe\x0e\x65\xbd\x65\x51\x3f\xa1\xa4\xab\x31\xe1\x09\x25\x3d\xca\xff\x79\x04\x54\x74\x84\x76\x29\xe8\x7f\xa2\x9c\x77\x83\xbf\x77\xb5\x6d\x43\xd3\x3e\xca\xb7\x2c\xe4\xe7\x2b\xe3\x7d\xf9\xdb\x93\x99\x06\x66\xa7\x55\xbe\xbd\xc5\x9b\x39\x9d\x52\xad\x1f\xfe\x78\xce\x6f\x1e\x74\x5a\xa3\x97\xbe\x48\xbf\xa5\x4b\x7d\x8d\xce\xa7\x07\x26\xa4\xb9\xa6\x2b\x81\xc5\x7f\x87\x9c\x99\x15\x67\xa4\xe2\x5f\x47\xd1\xcd\x4c\x2b\xc3\xa8\xbd\x90\x4b\xfa\x80\x73\x83\x3e\x66\x4b\xbc\xbe\xa1\xee\xf2\x71\x21\x40\x4a\x64\xde\x5c\xaa\x3e\x7f\xde\xcc\xf5\x2d\xc8\x56\xad\xda\x21\x08\xed\x58\x3f\xed\x18\xd2\x00\x71\x72\x54\x50\x65\xb7\x20\x00\xbc\x24\x82\x25\x1b\x25\xad\x47\x76\xb2\xa1\xf7\x91\xb6\x89\xab\xa3\xd1\x1b\x70\xa3\x34\x9c\x0f\x52\xae\xde\xed\x1a\xeb\x76\x04\x5f\x77\x0e\xa8\xaa\x61\xd9\xa1\x7c\x94\xa2\x41\x7f\xb5\x85\x2a\xd0\xcd\xc9\xfe\x90\x26\xc5\x34\x16\xe1\x7f\x1b\x9c\x9d\x26\xb1\xe7\x8f\xa2\x05\xf6\xa0\x0a\x67\xbf\xd1\xc5\xf9\x15\xa9\xbc\x2f\xee\x28\x4f\x0c\x78\x2b\x3c\x1f\xdc\x12\x4a\xee\xe6\x07\xf7\xd4\xf5\xbe\x68\x66\xec\x0c\xed\xbf\x97\x44\xcd\x9e\xb0\xa1\xa9\x54\xed\xe4\x3b\xb7\x6a\xdd\x57\x72\x9a\xde\xdf\xae\x16\x50\x5d\xc3\x79\x82\xd7\xee\x1f\xa8\x3c\xc3\x71\x20\xff\xd2\xb8\xd5\x86\x4c\x06\x82\xc5\x8c\xe3\x5b\xd8\xfe\x5d\xa0\xa0\xac\x75\x51\xb6\x77\x78\xbb\x5c\x69\x52\x7a\x75\xd3\x4f\x1c\x03\x47\x75\x43\x11\xed\xd9\x34\xaa\xb7\x09\xaf\xa9\x6c\x4e\xe5\xc5\x18\xc4\xbe\xeb\xf9\xdf\x01\x00\x00\xff\xff\xb6\x40\xa0\xfb\xf2\x15\x00\x00")
+var _templatesSchemaserializerGotmpl = []byte("\x1f\x8b\x08\x00\x00\x09\x6e\x88\x00\xff\xec\x59\x6b\x6f\xdb\x36\x17\xfe\x1c\xfd\x0a\xc6\xe8\xdb\x4a\x79\x3d\xb5\x1b\x86\x7d\xc8\x90\x01\x6d\xd3\x62\x19\xba\x64\x68\x7a\x01\xd6\x16\x05\x63\xd1\x09\x1b\x99\xf2\x48\x3a\x4e\xe6\xea\xbf\xef\xf0\x50\x96\x44\x89\x92\xed\xa4\xde\x05\xe8\x80\x35\xb6\x44\x9e\xcb\x73\x9e\x73\x21\xbd\x58\x90\x84\x8d\xb9\x60\x64\x30\xc9\x12\x96\x9e\x32\xc9\x69\xca\xff\x64\x72\x40\xf2\x3c\x78\xf8\x90\xfc\x4a\xa5\xba\xa0\xe9\x2f\xa7\x27\xc7\x64\x62\x3f\x2b\xa2\x2f\xb8\x22\xd9\xd9\x27\x36\xd2\x64\xce\xf5\x05\xa1\x49\xc2\x35\xcf\x04\x4d\xc9\x54\x66\x53\x26\x35\x67\x8a\x70\xa1\x33\x42\x09\xee\xb5\xab\x83\xf1\x4c\x8c\x48\xb8\x58\xc4\x2f\xd9\x88\xf1\x2b\x26\x8f\xe9\x84\xe5\x39\x01\x43\xa6\x54\x8d\x50\x37\x89\xcd\x43\xd0\x1f\xd5\xb5\x87\x11\x09\xdf\x7d\x38\xbb\xd1\x6c\x48\x98\x94\x99\x8c\xc8\x22\xd8\xc9\x66\x9a\xec\x1f\x90\x4f\x73\xc9\x35\x93\xf1\x5b\xfc\xb3\xc8\x83\x1d\x10\xe8\xe8\x00\x71\x71\x21\xed\x19\x55\x37\x28\xf1\x3e\xec\x8e\x82\x1d\xc9\xf4\x4c\x0a\x02\x5f\xe2\x27\x33\x9e\x26\x4f\x40\x87\x0a\xa3\x20\x0f\x0c\x00\xaf\xc5\xa4\x06\xc1\x4c\x6c\x0a\xc2\x58\x66\x13\x84\xa0\xd3\xf7\xbd\x0e\xe7\x1d\xcd\x61\x42\x35\x25\x16\x80\xc8\x02\x00\xfe\x13\xc0\x18\xfd\x4f\xd9\x35\xb8\xff\xc2\xfc\xbb\x38\x84\x95\xfb\xc4\xac\xef\xc0\xa1\x14\x5c\x21\xc1\x45\x05\x04\x17\xf1\x33\x23\xbf\x84\xa0\x81\xdb\x36\x79\xd0\x85\x45\x33\x74\x26\xee\x7b\x6e\xd4\x23\xc4\x03\xf6\x6b\x36\x99\xa6\x54\x03\xa7\xd5\xe8\x82\x4d\x68\x9d\xd4\xb1\xa1\x75\x23\xb0\xa5\x5f\xff\x48\x70\x4b\x97\x20\x92\x7b\xf5\x38\x76\xfa\x73\xc8\x54\xdb\x23\x58\xc7\x44\x62\x3e\xd7\x72\xda\xae\x7f\x43\xd3\x19\x73\x33\xdb\x58\x1b\x5e\x99\xe7\x3e\x05\xaf\x6e\xa6\xcc\x0a\x1e\x12\x2f\xd0\x15\xfd\x60\x37\x1f\x13\xf8\x12\x1f\xa9\xa7\x19\x88\x61\xd7\x27\x16\x35\x78\xf0\x6a\x06\xdf\xcd\x87\xc7\x25\x78\xbf\x55\xd8\x99\xe7\x29\xa7\x8a\x25\xb8\x79\xa6\x74\x36\x79\x9e\xc9\x09\xd5\xa0\x03\x74\x9f\x61\x9e\x1b\x76\xab\x39\x3d\xb7\xba\x11\x28\xb4\x3b\x32\xdc\x1f\xe3\x8a\xdd\x03\x22\x78\x8a\xe6\x10\x52\x70\x18\x2d\x84\x07\x39\xfc\x6f\x12\xfb\x25\x9d\x87\x20\x10\xd6\x45\x06\xa9\x54\x31\xb3\xdb\x98\x20\x25\xbd\x01\x6d\xc5\x22\x93\xfb\xe1\x83\x77\x0f\x22\x14\x36\x06\xc7\xf8\x90\x5c\x19\x23\x24\x15\xe7\x88\x56\x8c\x78\x3e\xbb\x9e\x4a\xa6\x14\x38\x45\x4c\xf9\xc2\xe5\x68\x11\x27\x3f\x91\x47\xe5\x83\x52\xbb\x15\x3c\x2c\x04\x5b\xc3\xcc\x7f\x46\xe0\x91\x48\xd8\xf5\x1b\x6a\x9c\x7e\x8e\xe9\xec\x89\x49\x2b\x88\xf1\x11\x2c\x51\x26\xe0\xa5\xea\x02\xae\x96\xc8\xf0\x0a\xe3\x18\xfd\xd8\x86\xcb\x85\xcc\xb1\xcd\xfe\xeb\x98\xff\x01\xcd\x5f\x02\x98\xe7\x46\x93\x09\xc9\x09\x90\x24\xcf\x8b\xc0\x54\x4c\x2c\x45\x83\xbe\x20\x5f\xc5\x50\x97\xd6\x05\x47\xdb\x49\x11\xae\x20\x6c\xd9\x1b\xb6\x48\xcd\x2b\x00\x16\xa2\x3f\x4b\xb5\x1b\xaa\x84\x49\x36\x1e\xb3\xe4\xb4\x61\x55\x00\x86\x60\xfd\x86\xe8\x40\x75\x35\x64\x84\x60\xc0\xa7\x93\xcb\xd0\xa6\x79\x15\x3d\x24\xfb\x4b\x46\x93\xb2\xea\x0f\xc9\x7d\xab\xcc\x17\xc0\x02\x61\xeb\x2c\x18\x7b\x3c\x4b\x53\x7a\x96\x9a\xe8\xc0\x22\x37\x54\xbf\x33\x99\xd9\x8f\x36\x10\xc3\x22\xe6\xa6\x7e\xf4\xc8\xb9\x5f\x6e\xb0\x66\x60\x1a\x05\xfe\x34\xaa\x21\xf3\xee\x43\x77\xa4\x4a\xee\x36\xb8\x6a\xfc\xee\x4f\x81\x46\xf1\x2b\x05\x01\x98\x87\x2c\xe5\x13\x9b\xbc\x60\xd4\x2e\x3c\x39\x52\xc5\x33\x20\x2e\xe2\x55\x98\x76\x00\x2d\xec\x92\x85\xeb\x58\x38\x24\x8f\x86\xe4\x87\xef\xa1\x15\x5a\x77\x1d\x29\xc8\xeb\xc0\x54\x09\xaf\xb6\xf9\xd5\xb0\x23\x27\xad\x9f\xa1\xe9\xba\xfe\x32\x06\xe2\x80\x99\xb6\x0d\xc3\xcb\xa8\x16\xeb\xa2\xaa\x95\x46\xd0\xe9\x14\xa2\x13\x2e\x63\x33\xbf\x42\x99\x22\x7e\x4b\x85\x06\xde\x4f\x28\xf6\xf1\x0a\x1f\x93\xc4\x15\x2f\xaa\xb8\x63\x26\x1f\x09\xf4\xb9\xa0\x8e\x4d\xd7\xbe\xcc\x6d\x0c\x8c\x77\xc8\x39\xe4\x65\x33\x18\x76\x73\xab\x81\xfb\xb9\xe7\x6b\xfd\x29\x1f\xb1\x15\xdb\xdb\xe9\xed\x13\xa4\xe5\x78\xd2\x6d\x48\xb3\x14\x56\xed\xb8\x07\xbb\x66\xc1\xdb\x12\x7a\x9e\x71\x61\x33\xfc\x56\x0a\xd8\x00\xc1\x1e\x59\x15\x86\x48\xc2\xb5\x20\x6c\x87\x25\xef\x1d\x85\x7c\x46\xac\xd8\xd1\x64\x10\x2c\x6a\xcd\x0a\xb0\x6b\x04\x83\x9d\x33\xe2\x61\x01\xb9\x4d\x3b\xdf\x68\xe8\x68\x0c\x1c\x9e\x61\x23\x0f\x1a\xb3\x41\x97\xa5\x3d\x53\x82\x61\x25\x48\xc6\x72\x44\x0e\xdc\xf9\xa1\xf6\xa2\x36\x42\x94\xa5\xca\x1a\x60\x3a\x4c\x6b\x92\x58\x85\x7a\x2b\x4c\x5f\x9b\xcb\x7f\xa6\xb9\x74\xc4\xd5\x57\x91\x8a\x40\x35\x0f\xa9\x1f\x3f\x0a\x18\x41\x14\xd8\xb0\x77\xc6\xb5\x62\x3a\x3e\x66\x73\x33\xfc\x89\xac\x3e\xa9\x99\xc3\x54\xc7\x7e\xd8\x83\x51\x5d\x4f\x86\xc9\x24\x51\x0c\x3e\xcb\xa1\x4c\xc4\xa7\x97\x7c\x1a\x46\xe5\x24\xeb\xf8\xba\x00\x5f\x5b\xe1\xc9\x8b\xf0\x5c\xb2\x9b\x62\xda\x7b\x2d\x14\x1d\x33\xa8\x3c\x5c\x9c\x87\x0e\x7c\x69\x26\xc2\x62\xa0\xbe\x67\x0e\x95\x2f\x18\xb2\x33\x74\xad\x8b\xec\x28\xed\xb1\x8f\xc0\xb0\xc8\xf5\xe8\x02\x95\xd9\x07\x20\xca\xd6\x8c\x7b\xdc\x10\x62\x68\x05\x1b\xa9\xb1\xd3\x2d\x70\xf1\x88\x2a\x2c\x2e\x53\x30\x4d\x8f\xc9\xe0\x7f\x7f\x0c\xec\xfa\x65\x65\xd8\xaf\x0e\x29\xf6\xb9\xa7\x0a\x1d\x54\x6f\x8b\xf9\xb2\xb6\xa9\x2f\x2a\xf1\x29\xd3\x26\x16\xd6\x52\x8c\x41\xef\x46\xa4\x43\xc7\xa6\x7a\xaf\x30\x3d\xa9\xa3\x47\xd6\xce\x5c\xbe\x15\x9d\xee\xf9\x57\xdf\xd2\xdd\x65\xa4\x3d\xc6\x2e\xeb\x02\xf9\x86\x7c\x5b\xba\xb4\x11\x2c\xb7\x13\xde\x0d\x5f\xe3\x90\xe9\x62\x81\x2f\xd7\x01\xcd\x2e\xbc\x33\x5e\xb7\x44\xa2\xe1\x23\x7e\x83\xb2\x44\xa1\xa2\x2d\x09\x5e\x64\x3a\xc8\x9c\x49\x05\x52\xc3\xa8\x76\x04\xae\x97\x81\x56\xf1\xc3\x3c\xca\x84\xe6\x62\xc6\x8a\xe2\xd9\x48\xca\x0d\x52\x72\xcd\x84\xac\x37\x71\xe7\x35\x22\xb7\x49\x8b\xb3\xa6\x80\xee\x60\x67\x07\x02\xdf\x2f\x78\xe8\x1b\x22\x7c\xfa\x4d\xeb\xf1\xde\x35\x78\x1b\x4f\xad\xf7\xad\x55\x6b\x7a\x54\x6f\xb5\xf2\x3c\x4d\x19\x95\x9e\x6d\x79\xb0\x59\xfd\x71\x46\x30\x6f\x61\xb9\x6d\x34\x3b\xd5\xb6\xa3\xbb\x5a\xb1\x37\xda\xeb\xd8\x7b\xe7\xe8\xb7\x22\xe0\x5e\xa1\xf6\x99\x40\x36\x35\xf7\x5f\x51\xbc\x2b\x66\xdd\x4e\x7c\x3f\x03\x9d\x12\xde\x81\x8e\x2d\xd0\x5f\x80\x77\xa5\xb2\x5e\xca\xb5\xd5\xad\x62\x9b\xcf\xc0\xbf\x8f\x68\x8e\xf6\x6e\x8e\xb5\x8d\xdc\x72\xaf\xf3\x11\xc7\x43\x89\xa0\xd1\xef\xbc\xdd\x2e\xef\x1f\xec\xf3\x75\x06\xfb\xbe\x73\x32\x4e\xcb\x5c\x2a\xfc\xbd\x4e\x4b\xc0\xe6\x23\x20\x89\x4f\x82\x9e\x16\xf9\x38\x4d\x4f\xc6\xc4\x5e\xa1\xee\xe2\xdd\xe8\x91\x32\x83\x44\xd8\xd3\x27\x22\xf2\xf9\xf3\x6a\xcc\x5f\x31\xd5\xe8\x02\xcb\x13\xf5\xae\xb5\xb3\xf7\x48\x6d\x97\x80\x03\x14\x58\x55\xfe\x10\x62\x2d\x72\x6e\x4f\x6d\x16\x74\xc7\xb0\xc3\x8c\x52\x71\x71\x70\x18\x98\xd5\x03\xd4\x5e\x27\x72\x3d\xed\xfd\x8d\x11\x7f\x37\x59\xff\x1e\xa2\x9c\x08\x90\x28\xde\x7b\x03\x9f\xf8\xbe\x70\xf4\xfc\xf8\xd0\x77\xb1\xd0\x75\xb5\xe0\x5c\x2e\x94\x3f\x56\x74\xe2\x1f\xd8\x97\xcb\xbb\x4c\xe7\x5a\xe9\x1e\xcd\x96\x67\xae\x14\xfe\x5a\xb2\x45\xf6\xdd\x7a\x33\xdb\xf6\x59\x69\x1d\x3b\xd7\x4b\x63\x1f\x99\x12\xf4\x7f\x52\x59\xbf\xbc\x20\xbb\x23\x81\x1b\x74\x7b\x3f\x28\x5d\x29\xcc\x7c\x3f\xd8\x1f\x44\x5f\x94\xe9\x1b\xbb\xf6\x35\x29\xb6\x9d\x14\x6e\xcd\x5e\x51\xf4\x2b\x88\xbe\x1b\x54\x93\xc7\x91\x3a\x9d\x9d\x99\x8b\x2e\x52\x5e\x69\x8b\x4c\x93\xf8\x67\xaa\x9e\xc0\xf9\xa6\x78\xe3\xc0\x9e\x70\x35\x92\x7c\xc2\x05\x7c\x4b\x9a\x97\xec\x4b\x0a\x34\x2e\x9d\xed\x7d\xb8\x7d\x5c\x49\xd2\xe6\xe9\xaa\xdf\x0b\x3a\xa6\x64\x47\x10\xf5\x2c\xea\x95\xeb\xba\xe7\x0a\xbb\xa0\xea\xb0\xcf\x45\xdf\x5d\xf7\x5f\x01\x00\x00\xff\xff\x15\x95\xbb\x6e\x9b\x23\x00\x00")
 
 func templatesSchemaserializerGotmplBytes() ([]byte, error) {
 	return bindataRead(
@@ -331,7 +331,7 @@ func templatesSchemaserializerGotmpl() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "templates/schemaserializer.gotmpl", size: 5618, mode: os.FileMode(420), modTime: time.Unix(1465223901, 0)}
+	info := bindataFileInfo{name: "templates/schemaserializer.gotmpl", size: 9115, mode: os.FileMode(420), modTime: time.Unix(1465223932, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
