+++ generator/model.go
@@ -241,7 +241,6 @@ func makeGenDefinitionHierarchy(name, pkg, container string, schema spec.Schema,
 	var defaultImports []string
 	if pg.GenSchema.HasValidations {
 		defaultImports = []string{
-			"github.com/go-openapi/errors",
 			"github.com/go-openapi/runtime",
 			"github.com/go-openapi/validate",
 		}
@@ -1271,44 +1270,6 @@ func (sg *schemaGenContext) makeGenSchema() error {
 		log.Printf("finished gen schema for %q\n", sg.Name)
 	}
 	return nil
-}
-
-var inEasyJSONMap map[string]string
-var outEasyJSONMap map[string]string
-
-func init() {
-	inEasyJSONMap = map[string]string{
-		"uint8":   "in.Uint8()",
-		"uint16":  "in.Uint16()",
-		"uint32":  "in.Uint32()",
-		"uint":    "in.Uint()",
-		"uint64":  "in.Uint64()",
-		"int8":    "in.Int8()",
-		"int16":   "in.Int16()",
-		"int32":   "in.Int32()",
-		"int":     "in.Int()",
-		"int64":   "in.Int64()",
-		"float32": "in.Float32()",
-		"float64": "in.Float64()",
-		"string":  "in.String()",
-		"bool":    "in.Bool()",
-	}
-	outEasyJSONMap = map[string]string{
-		"uint8":   "out.Uint8",
-		"uint16":  "out.Uint16",
-		"uint32":  "out.Uint32",
-		"uint":    "out.Uint",
-		"uint64":  "out.Uint64",
-		"int8":    "out.Int8",
-		"int16":   "out.Int16",
-		"int32":   "out.Int32",
-		"int":     "out.Int",
-		"int64":   "out.Int64",
-		"float32": "out.Float32",
-		"float64": "out.Float64",
-		"string":  "out.String",
-		"bool":    "out.Bool",
-	}
 }
 
 func (sg *schemaGenContext) jsonIn() string {
