+++ generator/templates/header.gotmpl
@@ -4,7 +4,9 @@ package {{.Package}}
 // Editing this file might prove futile when you re-run the swagger generate command
 
 {{if or .Imports .DefaultImports }}import (
+  errors  "github.com/go-openapi/errors"
   strfmt  "github.com/go-openapi/strfmt"
+  swag    "github.com/go-openapi/swag"
   jlexer  "github.com/mailru/easyjson/jlexer"
 	jwriter "github.com/mailru/easyjson/jwriter"
 	bitset  "github.com/willf/bitset"
